renaming PvtArgs to Args, PvtResult to Result
