+++ AndroidPvtLib/src/main/java/rs/arthu/androidpvt/lib/PvtResult.kt
@@ -1,0 +1,8 @@
+package rs.arthu.androidpvt.lib
+
+internal data class Result(
+    val testNumber: Int,
+    val timestamp: Long,
+    val interval: Long,
+    val reactionDelay: Long
+)
